remove v0 version of the database (#5698)
