+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 ### Breaking Changes
 
+- Removed support for version 0 of the database as it is no longer used by any active node.
+
 ### Additions and Improvements
 - EvmTool now executes the `execution-spec-tests` via the `t8n` and `b11r`. See the [README](ethereum/evmtool/README.md) in EvmTool for more instructions.
 - Improve lifecycle management of the transaction pool [#5634](https://github.com/hyperledger/besu/pull/5634)
